REQ - 21
REQ-21: The system shall allow variations in translations based on different contexts.
