+++ src/middleware/multerConfig.js
@@ -1,0 +1,4 @@
+// middleware/multerConfig.js
+const multer = require('multer');
+const upload = multer({ dest: 'uploads/' });
+module.exports = upload;
